setup: ensure source and target instance names are not the same

+++ cmd/setup/main.go
@@ -3,18 +3,18 @@ package main
 import (
 	"context"
 	"fmt"
+	"os"
+	"time"
+
 	"github.com/nais/cloudsql-migrator/internal/pkg/application"
 	"github.com/nais/cloudsql-migrator/internal/pkg/backup"
+	"github.com/nais/cloudsql-migrator/internal/pkg/common_main"
+	"github.com/nais/cloudsql-migrator/internal/pkg/config"
 	"github.com/nais/cloudsql-migrator/internal/pkg/database"
 	"github.com/nais/cloudsql-migrator/internal/pkg/instance"
 	"github.com/nais/cloudsql-migrator/internal/pkg/migration"
 	"github.com/nais/cloudsql-migrator/internal/pkg/netpol"
 	"github.com/nais/cloudsql-migrator/internal/pkg/resolved"
-	"os"
-	"time"
-
-	"github.com/nais/cloudsql-migrator/internal/pkg/common_main"
-	"github.com/nais/cloudsql-migrator/internal/pkg/config"
 	"github.com/sethvargo/go-envconfig"
 )
 
@@ -56,70 +56,75 @@ func main() {
 		os.Exit(5)
 	}
 
+	if source.Name == cfg.TargetInstance.Name {
+		mgr.Logger.Error("source and target instance cannot be the same")
+		os.Exit(6)
+	}
+
 	databaseName, err := resolved.ResolveDatabaseName(app)
 	if err != nil {
 		mgr.Logger.Error("failed to resolve database name", "error", err)
-		os.Exit(6)
+		os.Exit(7)
 	}
 
 	target, err := instance.CreateInstance(ctx, cfg, source, gcpProject, databaseName, mgr)
 	if err != nil {
 		mgr.Logger.Error("failed to create target instance", "error", err)
-		os.Exit(7)
+		os.Exit(8)
 	}
 
 	err = database.DeleteHelperTargetDatabase(ctx, cfg, target, databaseName, gcpProject, mgr)
 	if err != nil {
 		mgr.Logger.Error("failed to delete database from intended target instance", "error", err)
-		os.Exit(8)
+		os.Exit(9)
 	}
 
 	err = backup.CreateBackup(ctx, cfg, source.Name, gcpProject, mgr)
 	if err != nil {
 		mgr.Logger.Error("Failed to create backup", "error", err)
-		os.Exit(9)
+		os.Exit(10)
 	}
 
 	err = application.DisableCascadingDelete(ctx, cfg, mgr)
 	if err != nil {
 		mgr.Logger.Error("failed to disable cascading delete", "error", err)
-		os.Exit(10)
+		os.Exit(11)
 	}
 
 	err = netpol.CreateNetworkPolicy(ctx, cfg, source, target, mgr)
 	if err != nil {
 		mgr.Logger.Error("failed to create network policy", "error", err)
-		os.Exit(11)
+		os.Exit(12)
 	}
 
 	err = instance.PrepareSourceInstance(ctx, source, target, mgr)
 	if err != nil {
 		mgr.Logger.Error("failed to prepare source instance", "error", err)
-		os.Exit(12)
+		os.Exit(13)
 	}
 
 	err = database.PrepareSourceDatabase(ctx, cfg, source, databaseName, gcpProject, mgr)
 	if err != nil {
 		mgr.Logger.Error("failed to prepare source database", "error", err)
-		os.Exit(13)
+		os.Exit(14)
 	}
 
 	err = instance.PrepareTargetInstance(ctx, target, mgr)
 	if err != nil {
 		mgr.Logger.Error("failed to prepare target instance", "error", err)
-		os.Exit(14)
+		os.Exit(15)
 	}
 
 	err = database.PrepareTargetDatabase(ctx, cfg, target, gcpProject, mgr)
 	if err != nil {
 		mgr.Logger.Error("failed to prepare target database", "error", err)
-		os.Exit(15)
+		os.Exit(16)
 	}
 
 	err = migration.SetupMigration(ctx, cfg, gcpProject, source, target, mgr)
 	if err != nil {
 		mgr.Logger.Error("failed to setup migration", "error", err)
-		os.Exit(16)
+		os.Exit(17)
 	}
 
 	mgr.Logger.Info("setup completed")
